Fix the googleurl_unittests hack so iOS builds work again.
For this to work on iOS, the entire .app bundle needs to be dulpicated.

Further, the executable name can't be changed because the build process puts the
name of the executable in the Info.plist.

BUG=229660
TBR=brettw
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/14921002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ee126d275be184cba074f2638d122c0a3d774ef