Prevent URLs with invalid schemes from resolving as relative to data: URLs.
There is a missing bit of logic when the scheme is invalid that is
present in the case where the scheme is empty.  Either way, we can't
consider this a relative URL if the base scheme isn't heirarchical.

BUG=346132

Review URL: https://codereview.chromium.org/177093008

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45172e6e478a359eba9c05bc72dd0b4c241a65c8