GURL::Replacements methods accept a StringPiece instead of std::string&.
Previously, it was unsafe to pass a char* because it would be implicitly
converted to a string with a very short lifetime. Now, you can safely
pass a char* as long as the memory pointed to by the char* outlives the
Replacements (which is always true for string literals).

All existing usage of Replacements should continue to work as usual.

Updated every call to Replacements Set*Str methods that that could be
simplified by the new StringPiece API (for example, passing string
literals directly, instead of copying them into a string object).

Internally, renamed StdStringReplacements to StringPieceReplacements.

BUG=454274

Review URL: https://codereview.chromium.org/889463003

Cr-Original-Commit-Position: refs/heads/master@{#314771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77752c3e55c59747a62a362799ac6de5ff56fd45