Handle backslashes when resolving relative URLs on data URLs.
This was causing an assertion failure. The problem was that the path parsing was treating backslash as a path separator/terminator, but the relative URL resolving code did not. The reason is that the relative URL code assumed the input is canonicalized already. This is true in normal cases but not when the relative URLs are non-standard like "data:".

BUG=456148

Review URL: https://codereview.chromium.org/934733005

Cr-Original-Commit-Position: refs/heads/master@{#316732}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e66ce87932cfe85a770e11086b3f2054ab42839e