Add missing using declarations in cronet code.
Currently jni_generator_helper.h includes a using statement for
JavaParamRef and ScopedJavaLocalRef. To allow this to be removed (since
it shouldn't really be there in a header file), add a "using" statement
to cronet-specific files that do this (already removed for other android
configs).

BUG=568574

Review-Url: https://codereview.chromium.org/2215903002
Cr-Original-Commit-Position: refs/heads/master@{#410032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc8890f04a40ef61b0f3b02e08a0fe6980aa81f5