Add url::Origin::GetURL() to convert Origins to URLs without reparsing
The previous canonical way to convert an Origin to a GURL was
to call GURL(origin.Serialize()). This is expensive and often in the
critical path of user perceived page loads. One such caller has
been converted to the new method, and the rest will be converted
in a followup patch.

BUG=651554

Review-Url: https://codereview.chromium.org/2378323003
Cr-Original-Commit-Position: refs/heads/master@{#422613}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 048bee100cbbc999219901a759fc86abc67c726d