Add operator==(const GURL&, const StringPiece&) to gurl.h.
This patch moves GURL::operator==(const GURL&) to a free-floating function,
and adds a new function to allow comparisons with StringPiece specs.

The new equality operator should be used when the spec of the URL is known
apriori, and parsing a GURL is not necessary for a comparison. This is
specifically designed to replace the case where we compare:
<url> = GURL(kUrlStringConstant).

One such case is replaced by this patch, and the rest will be modified in
a followup.

Note: this patch requires changes to instances of
base::Bind(&GURL::operator==...). They have been edited to use static_cast
to let Bind know that the operator== acts on GURLs.

BUG=655467

Review-Url: https://codereview.chromium.org/2421383003
Cr-Original-Commit-Position: refs/heads/master@{#425863}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ebeca8e21727b975523416e9abaa2a4024ac36f5