[url] Avoid scanning for whitespace twice during ResolveRelative
For calls to url::ResolveRelative, if the input is not in fact relative,
we call RemoveURLWhitespace twice on the same buffer. This is not necessary.

On a test of loading 100 identical ~1mb data URL images, this patch speeds
up the renderer by ~8% on Linux.

BUG=657978

Review-Url: https://codereview.chromium.org/2540893004
Cr-Original-Commit-Position: refs/heads/master@{#435604}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c64537201a63fc8436fcedd8edab32be3669405b