Do not sanitize about:blank/#foo & about:blank?foo
This CL ensures that the browser will not sanitize these two types of
URLs. This fixes an issue with PlzNavigate where we could not go back to
about:blank/#foo due to some discrepancy between the url in the
FrameNavigationEntry & the url stored in the PageState of the same
entry.

BUG=575210

Review-Url: https://codereview.chromium.org/2644133002
Cr-Original-Commit-Position: refs/heads/master@{#445525}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eff9252e95930e904f389f7f902bd9c815243f82