Share schemes needed for CSP between the browser and the renderer.
This CL is similar to this one: crrev.com/2623353002/
There is a list of scheme that bypasses the Content-Security-Policy.
With this CL, this list is provided by content embedders instead of
blink embedders. It will be used for checking CSP in the browser process.

BUG=685074

Review-Url: https://codereview.chromium.org/2679383003
Cr-Original-Commit-Position: refs/heads/master@{#449261}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb73e4330ecb59905157b11ce20ca29345d9635e