[url] For empty hosts, just trust GURL to make the invalid/valid decision
The SchemeHostPort always considers tuples with SCHEME_WITHOUT_PORT to
have empty hosts without being invalid (to cater to cases like file
schemes which can have hosts or not).

This led to a mismatch in SchemeHostPost::GetURL where we always
considered URLs with empty hosts to be valid GURLs, but this is not an
assumption that holds true. This CL just punts the valid decision in this
case to GURL for re-parsing, since it should be a rare occurance.

Bug: 820194
Change-Id: I3f4441d8c71a9b9500c71dbe98f0769f42ad13b5
Reviewed-on: https://chromium-review.googlesource.com/956424
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#542527}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60f0c536591d5b64febafb5f3c5112ca33c691a4